html2/requirements.txt
Flask
flask-cors
mysql-connector-python
python-dotenv
numpy
Werkzeug
gunicorn
numpy==2.1.3
opencv-python==4.12.0.88
mediapipe==0.10.14
tensorflow==2.19.1
scikit-learn==1.7.2
tqdm==4.67.1
mysql-connector-python==9.4.0
absl-py==2.3.1
astunparse==1.6.3
attrs==25.4.0
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.1
h5py==3.15.0
joblib==1.5.2
keras==3.11.3
libclang==18.1.1
ml-dtypes==0.5.3
opt-einsum==3.4.0
protobuf==4.25.8
requests==2.32.5
scipy==1.16.2
six==1.17.0
sounddevice==0.5.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
typing-extensions==4.15.0
wrapt==1.17.3
colorama==0.4.6
opencv-contrib-python==4.12.0.88
matplotlib==3.10.7
packaging==25.0
werkzeug==3.1.3
CFFI==2.0.0
rich==14.2.0
jax==0.7.2
jaxlib==0.7.2
